Add project contacts to the README
Adding some text like this to the repo I think should let us check off
the Project has a suitable governance model box in the checklist.

Having placed the repo in the PRG lab github repo as opposed to it being
in a personal repo of Yang's of Gavin's gives us a check for the silver
plans in the even an admin is no longer available.

Once on Zenodo we can check off the archived on other platforms box and
get a silver overall in this category.